Refactored folder structure and code
