+++ src/components/Editor/Menu/Fixed/index.jsx
@@ -13,7 +13,6 @@ import MediaUploader from "components/Editor/MediaUploader";
 import LinkAddPopOver from "./components/LinkAddPopOver";
 import MoreMenu from "./components/MoreMenu";
 import { MENU_ELEMENTS } from "./constants";
-// import HighlightDropdown from "./HighlightDropdown";
 import useEditorState from "./hooks/useEditorState";
 import { reGroupMenuItems, buildMenuOptions } from "./utils";
 
@@ -181,7 +180,6 @@ const Fixed = ({
             );
           })
         )}
-        {/* <HighlightDropdown /> */}
         {isNotEmpty(moreMenuItems) && (
           <MoreMenu {...{ editor }} groups={moreMenuItems} />
         )}
